Split Uniform & Attribute code into separate file
